Version 1.2.15: Refactored: Split to base_env.py and env.py
